+++ runtime/Apps/image_comparison/libfpng.dll
